Update keywords.py
This avoids a KeyError when, for whatever reason, two keywords occur right after another:

```python
from summa import keywords
keywords.keywords("Rabbit populations known to be plentiful, large, and diverse in the area. Adjacent to the site, a number number well over a thousand. The number of these rabbit populations has diminished in recent years, and perhaps we have become number to a number of their numbers numbering fewer.")
```

This results in the following for me:

```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "textrank/summa/keywords.py", line 212, in keywords
    combined_keywords = _get_combined_keywords(keywords, text.split())
  File "textrank/summa/summa/keywords.py", line 158, in _get_combined_keywords
    for keyword in combined_word: _keywords.pop(keyword)
KeyError: 'number'
```