Fix flaky test that calculated wrongly the ratio between keywords lenght. (#72)
This was caused when calculating keywords using stopwords. As there are
more words that are filtering, the list of concepts gets smaller and the error
gets bigger.

Because of this, one of every five runs failed because the ratio between runs
was slightly different than two.

This was fixed expecting that the ratio of the runs is double with 5% tolerance
instead of up to one decimal place.